# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: MLAKP Shared Expense API
  version: 0.1.0
  description: HTTP API contract for the shared expense and debt tracking backend.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
            example:
              refresh_token: opaque-refresh-token
      responses:
        "200":
          description: Refresh token accepted and rotated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
              example:
                success: true
                data:
                  access_token: eyJzdWIiOiIxMjMifQ.signature
                  refresh_token: rotated-refresh-token
                  token_type: Bearer
                  expires_at: "2026-05-02T12:15:00Z"
        "400":
          description: Invalid JSON.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: invalid_json
                  message: Request body must be valid JSON
        "401":
          description: Refresh token is invalid, expired, rotated, or revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: invalid_refresh_token
                  message: Refresh token is invalid or expired
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: HMAC-SHA256
  responses:
...
    RequestBodyTooLarge:
      description: Request body exceeds the maximum accepted size.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: request_body_too_large
              message: Request body must be 1 MiB or smaller
    TooManyRequests:
      description: Request rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: rate_limited
              message: Too many requests. Please retry later
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: internal_error
              message: Internal server error
  schemas:
...
    ErrorDetail:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          $ref: "#/components/schemas/ErrorDetail"
...
    RefreshTokenRequest:
      type: object
      additionalProperties: false
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          minLength: 1
          writeOnly: true
...
    RefreshTokenResponse:
      type: object
      additionalProperties: false
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          enum:
            - true
        data:
          $ref: "#/components/schemas/RefreshTokenData"
...