# internal/httpapi/handlers/auth.go
package handlers

import (
	"encoding/json"
	"errors"
	"io"
	"net/http"
	"time"

	"mlakp-backend/internal/auth"
	"mlakp-backend/internal/httpapi/middleware"
	"mlakp-backend/internal/httpapi/response"
	"mlakp-backend/internal/sessions"
	"mlakp-backend/internal/users"
)

const maxRequestBodyBytes int64 = 1 << 20

var errRequestBodyTooLarge = errors.New("request body too large")

type AuthHandler struct {
	users        *users.Service
	tokenManager *auth.TokenManager
	sessions     *sessions.Service
}

type authUserResponse struct {
	ID       string `json:"id"`
	Name     string `json:"name"`
	Username string `json:"username"`
	Email    string `json:"email"`
}

type tokenResponse struct {
	AccessToken          string           `json:"access_token"`
	RefreshToken         string           `json:"refresh_token"`
	TokenType            string           `json:"token_type"`
	ExpiresAt            string           `json:"expires_at"`
	User                 authUserResponse `json:"user"`
	VerificationWarning  string           `json:"verification_warning,omitempty"`
	VerificationDeadline *time.Time       `json:"verification_deadline,omitempty"`
	VerificationStatus   string           `json:"verification_status,omitempty"`
}

type refreshResponse struct {
	AccessToken  string `json:"access_token"`
	RefreshToken string `json:"refresh_token"`
	TokenType    string `json:"token_type"`
	ExpiresAt    string `json:"expires_at"`
}

func NewAuthHandler(users *users.Service, tokenManager *auth.TokenManager, sessions *sessions.Service) *AuthHandler {
	return &AuthHandler{
		users:        users,
		tokenManager: tokenManager,
		sessions:     sessions,
	}
}

func (h *AuthHandler) Register(w http.ResponseWriter, r *http.Request) {
	var request struct {
		Name     string `json:"name"`
		Username string `json:"username"`
		Email    string `json:"email"`
		Password string `json:"password"`
	}
	if err := decodeJSON(r, &request); err != nil {
		writeDecodeError(w, err)
		return
	}

	user, err := h.users.Register(r.Context(), request.Name, request.Username, request.Email, request.Password)
	if err != nil {
		writeUserError(w, err)
		return
	}

	session, refreshToken, err := h.sessions.Create(r.Context(), user.ID)
	if err != nil {
		response.Error(w, http.StatusInternalServerError, "internal_error", "Internal server error")
		return
	}

	accessToken, expiresAt, err := h.tokenManager.IssueAccessToken(r.Context(), user.ID, session.ID)
	if err != nil {
		response.Error(w, http.StatusInternalServerError, "internal_error", "Internal server error")
		return
	}

	writeAuthNoStoreHeaders(w)
	resp := tokenResponse{
		AccessToken:  accessToken,
		RefreshToken: refreshToken,
		TokenType:    "Bearer",
		ExpiresAt:    expiresAt.Format(timeFormatRFC3339),
		User:         toAuthUserResponse(user),
	}
	addVerificationFields(&resp, h.users.GetVerificationStatus(&user))
	response.Success(w, http.StatusCreated, resp)
}

func (h *AuthHandler) Login(w http.ResponseWriter, r *http.Request) {
	var request struct {
		Email    string `json:"email"`
		Password string `json:"password"`
	}
	if err := decodeJSON(r, &request); err != nil {
		writeDecodeError(w, err)
		return
	}

	user, err := h.users.Authenticate(r.Context(), request.Email, request.Password)
	if err != nil {
		if errors.Is(err, users.ErrNotFound) {
			response.Error(w, http.StatusUnauthorized, "invalid_credentials", "Email or password is incorrect")
			return
		}
		writeUserError(w, err)
		return
	}

	verificationStatus := h.users.GetVerificationStatus(&user)
	if verificationStatus.Status == "expired" {
		response.Error(w, http.StatusForbidden, "email_verification_required", "Please verify your email before logging in")
		return
	}

	session, refreshToken, err := h.sessions.Create(r.Context(), user.ID)
	if err != nil {
		response.Error(w, http.StatusInternalServerError, "internal_error", "Internal server error")
		return
	}

	accessToken, expiresAt, err := h.tokenManager.IssueAccessToken(r.Context(), user.ID, session.ID)
	if err != nil {
		response.Error(w, http.StatusInternalServerError, "internal_error", "Internal server error")
		return
	}

	resp := tokenResponse{
		AccessToken:  accessToken,
		RefreshToken: refreshToken,
		TokenType:    "Bearer",
		ExpiresAt:    expiresAt.Format(timeFormatRFC3339),
		User:         toAuthUserResponse(user),
	}
	addVerificationFields(&resp, verificationStatus)

	writeAuthNoStoreHeaders(w)
	response.Success(w, http.StatusOK, resp)
}

func (h *AuthHandler) Refresh(w http.ResponseWriter, r *http.Request) {
	var request struct {
		RefreshToken string `json:"refresh_token"`
	}
	if err := decodeJSON(r, &request); err != nil {
		writeDecodeError(w, err)
		return
	}

	session, refreshToken, err := h.sessions.Refresh(r.Context(), request.RefreshToken)
	if err != nil {
		if errors.Is(err, sessions.ErrInvalidRefreshToken) {
			response.Error(w, http.StatusUnauthorized, "invalid_refresh_token", "Refresh token is invalid or expired")
			return
		}
		response.Error(w, http.StatusInternalServerError, "internal_error", "Internal server error")
		return
	}

	user, err := h.users.GetByID(r.Context(), session.UserID)
	if err != nil {
		response.Error(w, http.StatusInternalServerError, "internal_error", "Internal server error")
		return
	}
	if h.users.GetVerificationStatus(&user).Status == "expired" {
		response.Error(w, http.StatusForbidden, "email_verification_required", "Please verify your email before refreshing your session")
		return
	}

	accessToken, expiresAt, err := h.tokenManager.IssueAccessToken(r.Context(), session.UserID, session.ID)
	if err != nil {
		response.Error(w, http.StatusInternalServerError, "internal_error", "Internal server error")
		return
	}

	writeAuthNoStoreHeaders(w)
	response.Success(w, http.StatusOK, refreshResponse{
		AccessToken:  accessToken,
		RefreshToken: refreshToken,
		TokenType:    "Bearer",
		ExpiresAt:    expiresAt.Format(timeFormatRFC3339),
	})
}

func (h *AuthHandler) Logout(w http.ResponseWriter, r *http.Request) {
	sessionID, ok := middleware.SessionIDFromContext(r.Context())
	if !ok {
		response.Error(w, http.StatusUnauthorized, "unauthenticated", "Authentication is required")
		return
	}
	if err := h.sessions.Revoke(r.Context(), sessionID); err != nil {
		response.Error(w, http.StatusInternalServerError, "internal_error", "Internal server error")
		return
	}

	writeAuthNoStoreHeaders(w)
	response.Success(w, http.StatusOK, nil)
}

func writeUserError(w http.ResponseWriter, err error) {
	switch {
	case errors.Is(err, users.ErrInvalidName):
		response.Error(w, http.StatusBadRequest, "invalid_name", "Name must be between 1 and 120 characters")
	case errors.Is(err, users.ErrInvalidUsername):
		response.Error(w, http.StatusBadRequest, "invalid_username", "Username must be 3 to 30 lowercase letters, numbers, or underscores")
	case errors.Is(err, users.ErrInvalidEmail):
		response.Error(w, http.StatusBadRequest, "invalid_email", "Email is invalid")
	case errors.Is(err, users.ErrInvalidPassword):
		response.Error(w, http.StatusBadRequest, "invalid_password", "Password must be at least 8 characters")
	case errors.Is(err, users.ErrUsernameConflict):
		response.Error(w, http.StatusConflict, "username_already_registered", "Username is already registered")
	case errors.Is(err, users.ErrEmailConflict):
		response.Error(w, http.StatusConflict, "email_already_registered", "Email is already registered")
	case errors.Is(err, users.ErrNotFound):
		response.Error(w, http.StatusNotFound, "user_not_found", "User was not found")
	default:
		response.Error(w, http.StatusInternalServerError, "internal_error", "Internal server error")
	}
}

func decodeJSON(r *http.Request, destination any) error {
	// Read one byte beyond the allowed size so oversized bodies become an
	// explicit 413 instead of looking like truncated or malformed JSON.
	reader := &io.LimitedReader{R: r.Body, N: maxRequestBodyBytes + 1}
	decoder := json.NewDecoder(reader)
	decoder.DisallowUnknownFields()

	if err := decoder.Decode(destination); err != nil {
		if reader.N <= 0 {
			return errRequestBodyTooLarge
		}
		return err
	}
	// A second decode must hit EOF; otherwise the body contains multiple JSON
	// values such as "{} {}", which the API contract does not allow.
	if err := decoder.Decode(&struct{}{}); err != io.EOF {
		if reader.N <= 0 {
			return errRequestBodyTooLarge
		}
		return err
	}
	if reader.N <= 0 {
		return errRequestBodyTooLarge
	}

	return nil
}

func writeDecodeError(w http.ResponseWriter, err error) {
	// Keep malformed JSON and oversized bodies separate so clients can retry
	// with a smaller payload instead of guessing why decoding failed.
	if errors.Is(err, errRequestBodyTooLarge) {
		response.Error(w, http.StatusRequestEntityTooLarge, "request_body_too_large", "Request body must be 1 MiB or smaller")
		return
	}

	response.Error(w, http.StatusBadRequest, "invalid_json", "Request body must be valid JSON")
}

func writeAuthNoStoreHeaders(w http.ResponseWriter) {
	w.Header().Set("Cache-Control", "no-store")
	w.Header().Set("Pragma", "no-cache")
}

func toAuthUserResponse(user users.User) authUserResponse {
	return authUserResponse{
		ID:       user.ID,
		Name:     user.Name,
		Username: user.Username,
		Email:    user.Email,
	}
}

func addVerificationFields(resp *tokenResponse, status users.VerificationStatus) {
	if status.IsVerified {
		return
	}
	resp.VerificationStatus = status.Status
	resp.VerificationDeadline = status.Deadline
	if status.Status == "pending_grace_period" {
		resp.VerificationWarning = "Please verify your email to keep account access after the grace period."
	}
}

const timeFormatRFC3339 = "2006-01-02T15:04:05Z07:00"